# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PU Transit API
servers:
  - url: /api
    description: Base API path
security:
  - FirebaseBearer: []
...
paths:
...
  /memberships/{uid}:
    patch:
      operationId: updateMembership
      tags: [memberships]
      summary: Update a membership
      parameters:
        - $ref: "#/components/parameters/Uid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MembershipUpdate"
      responses:
        "200":
          description: Updated membership
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  securitySchemes:
    FirebaseBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Uid:
      name: uid
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Unauthorized:
      description: Missing or invalid Firebase JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Firebase email, membership, role, or assignment is not allowed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ServiceUnavailable:
      description: Firebase or migration service unavailable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
...
    MemberRole:
      type: string
      enum: [student, staff, driver, admin]
    RequestedRole:
      type: string
      enum: [driver, admin]
    Member:
      type: object
      additionalProperties: false
      required: [uid, email, role, status, active, assignedBusId, requestedRole, expiresAt, root, createdAt, updatedAt]
      properties:
        uid:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/MemberRole"
        status:
          type: string
          enum: [pending, approved, rejected, suspended]
        active:
          type: boolean
        assignedBusId:
          type: string
        requestedRole:
          description: Role the member asked for at sign-up; cleared when an admin grants or dismisses it
          oneOf:
            - $ref: "#/components/schemas/RequestedRole"
            - type: "null"
        expiresAt:
          description: IDN-01 — UTC ms after which Rules and the API refuse protected access even though status/active are unchanged; null means no expiry
          oneOf:
            - type: integer
              format: int64
            - type: "null"
        root:
          description: True for the owner account, which no other admin can change
          type: boolean
        createdAt:
          type: integer
          format: int64
        updatedAt:
          type: integer
          format: int64
...
    MembershipUpdate:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        role:
          $ref: "#/components/schemas/MemberRole"
        status:
          type: string
          enum: [pending, approved, rejected, suspended]
        active:
          type: boolean
        assignedBusId:
          type: string
          maxLength: 128
        requestedRole:
          description: Only null is accepted, to dismiss a pending role request
          type: "null"
        expiresAt:
          description: IDN-01 — set or extend the access end (UTC ms, must be positive); null clears it. An admin cannot give themself a past expiry.
          oneOf:
            - type: integer
              format: int64
              minimum: 1
              maximum: 8640000000000000
            - type: "null"
...